fix some more obvious 64bit issues
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1162 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ RenderDevice.cpp
@@ -431,7 +431,7 @@ void RenderDevice::CreatePixelShader( const char* shader )
 			gShader = NULL;
 		}
 		ID3DXBuffer *tmp;
-		CHECKD3D(D3DXCompileShader( shader, strlen(shader), 0, 0, "ps_main", "ps_2_a", D3DXSHADER_OPTIMIZATION_LEVEL3|D3DXSHADER_PREFER_FLOW_CONTROL, &tmp, 0, 0 ));
+		CHECKD3D(D3DXCompileShader( shader, (UINT)strlen(shader), 0, 0, "ps_main", "ps_2_a", D3DXSHADER_OPTIMIZATION_LEVEL3|D3DXSHADER_PREFER_FLOW_CONTROL, &tmp, 0, 0 ));
 		CHECKD3D(m_pD3DDevice->CreatePixelShader( (DWORD*)tmp->GetBufferPointer(), &gShader ));
 		shader_cache = shader;
 	}
